moved fish.js, FishDataProvider.js, and FishList.js to fish folder under scripts.
